% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_error.R
\name{calculate.error}
\alias{calculate.error}
\title{Prediction error calculation}
\usage{
calculate.error(rf, true, test.set = NULL)
}
\arguments{
\item{rf}{Object of class \code{\link[ranger]{ranger}}.}

\item{true}{Vector with true value for each sample.}

\item{test.set}{Matrix or data.frame with predictor variables of test set
with variables in columns and samples in rows (Note: missing values are not
allowed).}
}
\value{
numeric vector with two elements for regression and probability
estimation (rmse, rsq) and five elements for classification (acc, err, mcc,
sens, spec)
}
\description{
Calculates prediction errors by comparing predicted values with
the true outcome values. For regression and probability mode, it will give
mean squared error (mse) and pseudo R-squared (rsq). For classification mode,
overall accuracy (acc), overall error (err), Matthews correlation coefficient
(mcc), sensitivity (sens) and specificity (spec) are returned.
}
