% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pw_rf_hunt.R
\name{pw.rf.hunt}
\alias{pw.rf.hunt}
\title{Pathway-guided random forest based on pathway importance score}
\usage{
pw.rf.hunt(x = NULL, y = NULL, info.pw = NULL, type = "regression",
  importance = "impurity_corrected", min.no.var = 5,
  not.assoc.pw = TRUE, ...)
}
\arguments{
\item{x}{Matrix or data.frame of predictor variables with variables in
columns and samples in rows (Note: missing values are not allowed).}

\item{y}{Vector with values of outcome variable (Note: will be converted
to factor if classification mode is used).}

\item{info.pw}{Matrix or data.frame assigning variables to pathways with
variables in rows and pathways in columns (0: variable does not belong to
pathway, 1: variable belongs to pathway).}

\item{type}{Type of random forest ('regression' or 'classification'). Default
is 'regression'.}

\item{importance}{Variable importance mode ('none', 'impurity',
'impurity_corrected' or 'permutation'). Default is 'impurity_corrected'.}

\item{min.no.var}{Minimal number of variables per pathway (Note: pathways
with less variables are removed or moved into 'not_associated_genes_pw').
Default is 5.}

\item{not.assoc.pw}{Combines all variables not associated to any pathway
or pathways with less then <min.no.var> into pathway
'not_associated_genes_pw'. Default is TRUE.}

\item{...}{Additional parameters used in the \code{\link{wrapper.rf}}
function.}
}
\value{
List with the following components:
\itemize{
\item \code{results.pw} data.frame with information about each pathway
\itemize{
\item \code{id} pathway identifier
\item \code{no.var} number of variables in pathway
\item \code{x.score} the pathway enrichment score of the pathway
\item \code{z.score} the standardized pathway enrichment score
\item \code{pval} P value based on standard normal distribution
\item \code{pval.adj} Bonferroni adjusted P value
\item \code{selected} 0 or 1 denoting if pathway is selected
}
\item \code{pw.sel} vector with ids of selected pathways
\item \code{info.pw} actual used variables to pathways assignment
\item \code{importance.genes} gene specific importances
}
}
\description{
Trains a random forest (RF) using all predictor variables and
estimates variable specific importance. A pathway importance score
(enrichment score) is calculated using the mean of the importance scores of
the pathway members. A standardized score is then computed for each pathway
which is used to obtain P values under the assumption that the standardized
scores are standardnormally distributed. For more details about random
forests parameters see \code{\link{wrapper.rf}} and for more details about
the method see Chen et al. (2012).
}
\examples{
# define pathway parameters
info.par = data.frame(pw = paste0("pw", 1:3),
                      no.genes = rep(20, 3),
                      prop.de = c(0, 0.5, 1),
                      cor = rep(0.6, 3),
                      stringsAsFactors = FALSE)

# simulate 30 samples (including 15 cases)
sim.data.l = sim.data.study.1(info.par = info.par,
                              no.samples = 30,
                              no.cases = 15,
                              seed = 42,
                              gamma.range = c(0.5, 1.5))

res.hunt = pw.rf.hunt(x = sim.data.l$data.train[, -1],
                      y = sim.data.l$data.train[, 1],
                      info.pw = sim.data.l$info.pw,
                      type = "classification",
                      not.assoc.pw = FALSE)

}
\references{
Chen, X. & Ishwaran, H. (2012) Pathway hunting by random survival
forests. Bioinformatics 29:99–105.
\url{https://doi.org/10.1093/bioinformatics/bts643}.
}
