% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pw_rf_synthetic_features.R
\name{pw.rf.synthetic.features}
\alias{pw.rf.synthetic.features}
\title{Pathway-guided random forest based on synthetic features}
\usage{
pw.rf.synthetic.features(x = NULL, y = NULL, info.pw = NULL,
  type = "regression", importance = "impurity_corrected", min.no.var = 5,
  not.assoc.pw = TRUE, ...)
}
\arguments{
\item{x}{Matrix or data.frame of predictor variables with variables in
columns and samples in rows (Note: missing values are not allowed).}

\item{y}{Vector with values of outcome variable (Note: will be converted
to factor if classification mode is used).}

\item{info.pw}{Matrix or data.frame assigning variables to pathways with
variables in rows and pathways in columns (0: variable does not belong to
pathway, 1: variable belongs to pathway).}

\item{type}{Type of random forest ('regression' or 'classification'). Default
is 'regression'.}

\item{importance}{Variable importance mode ('none', 'impurity',
'impurity_corrected' or 'permutation'). Default is 'impurity_corrected'.}

\item{min.no.var}{Minimal number of variables per pathway (Note: pathways
with less variables are removed or moved into 'not_associated_genes_pw').
Default is 5.}

\item{not.assoc.pw}{Combines all variables not associated to any pathway
or pathways with less then <min.no.var> into pathway
'not_associated_genes_pw'. Default is TRUE.}

\item{...}{Additional parameters used in the ranger or the associated
wrapper function.}
}
\value{
List with the following components:
\itemize{
\item \code{results.pw} data.frame with information about each pathway
\itemize{
\item \code{id} pathway identifier
\item \code{no.var} number of variables in pathway
\item \code{run.x} importances of attributes gathered in each importance
source run
\item \code{decision} Boruta decision (Confirmed, Rejected or Tentative)
\item \code{selected} 0 or 1 denoting if pw is selected
}
\item \code{pw.sel} vector with ids of selected pathways
\item \code{info.pw} actual used variables to pathways assignment
}
}
\description{
Trains a separate random forest (RF) for each pathway and
determines predictions for each of them. These predictions (called synthetic
features) are subsequently analyzed with the variable selection method
\code{\link[Boruta]{Boruta}} (Kursa et al. 2010) to select
important synthetic features and, hence, important pathways. For
more details about random forests parameters see \code{\link{wrapper.rf}} and
for more details about the method see Pan et al. (2014).
}
\details{
In the original implementation (Pan et al. 2014) the RF based on the
synthetic features is only evaluated regarding its prediction performance.
We additionally implemented the Boruta method to select important synthetic
features, i.e. pathways.
}
\references{
\itemize{
\item Pan, Q, et al. (2014) A system‐level pathway‐phenotype association
  analysis using synthetic feature random forest. Genetic Epi 38: 209-219.
  \url{https://dx.doi.org/10.1002/gepi.21794}.
\item Kursa, MB, & Rudnicki, WR. (2010) Feature selection with the
  Boruta package. J Stat Softw 36: 1-13.
  \url{http://dx.doi.org/10.18637/jss.v036.i11}.
  }

#' @examples
# define pathway parameters
info.par = data.frame(pw = paste0("pw", 1:3),
                      no.genes = rep(20, 3),
                      prop.de = c(0, 0.5, 1),
                      cor = rep(0.6, 3),
                      stringsAsFactors = FALSE)

# simulate 30 samples (including 15 cases)
sim.data.l = sim.data.study.1(info.par = info.par,
                              no.samples = 30,
                              no.cases = 15,
                              seed = 42,
                              gamma.range = c(0.5, 1.5))

res.sf = pw.rf.synthetic.features(x = sim.data.l$data.train[, -1],
                                  y = sim.data.l$data.train[, 1],
                                  info.pw = sim.data.l$info.pw,
                                  type = "classification",
                                  not.assoc.pw = FALSE)
}
