% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_data_study_1.R
\name{sim.data.study.1}
\alias{sim.data.study.1}
\title{Simulate data sets for simulation study 1}
\usage{
sim.data.study.1(info.par, no.samples, no.cases, seed, dd1 = 0.5, gamma.range)
}
\arguments{
\item{info.par}{Information about pathways as data.frame with columns
no.genes (number of genes in pathway), prop.de (proportion of genes to be
simulated as differentially expressed) and cor (pairwise correlation between
genes).}

\item{no.samples}{Total number of individuals to simulate.}

\item{no.cases}{Number of cases to simulate.}

\item{seed}{Random seed. Default is NULL, which generates the seed from R.}

\item{dd1}{Probability of up regulation.}

\item{gamma.range}{Range of absolute effect sizes.}
}
\value{
List with the following components:
\itemize{
\item \code{data.train} data.frame of training data with samples in rows and
variables in columns; the first column, called y, denotes the case-control
status (0: control, 1: case)
\item \code{data.test} data.frame of test data with samples in rows and
variables in columns; the first column, called y, denotes the case-control
status (0: control, 1: case)
\item \code{info.pw} data.frame assigning variables to pathways with
variables in rows and pathways in columns (0: variable does not belong to
pathway, 1: variable belongs to pathway).
}
}
\description{
Simulates pathways with specified number of genes, amount of
differential expression and pairwise correlation.
}
\details{
This function uses a modifed version of R code provided by Poisson
et al. (2011) as supplementary material.
}
\examples{
# define pathway parameters
info.par = data.frame(pw = paste0("pw", 1:3),
                      no.genes = rep(20, 3),
                      prop.de = c(0, 0.5, 1),
                      cor = rep(0.6, 3),
                      stringsAsFactors = FALSE)

# simulate 30 samples (including 15 cases)
sim.data.l = sim.data.study.1(info.par = info.par,
                              no.samples = 30,
                              no.cases = 15,
                              seed = 42,
                              gamma.range = c(0.5, 1.5))

}
\references{
Poisson LM, Taylor JM, Ghosh D. (2011) Integrative set enrichment testing
for multiple omics platforms. BMC Bioinformatics. 12:459.
\url{https://doi.org/10.1186/1471-2105-12-459}.
}
