% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_data_study_2.R
\name{sim.data.study.2}
\alias{sim.data.study.2}
\title{Simulate data sets for simulation study 2}
\usage{
sim.data.study.2(mvn = NULL, cov.matrix = NULL, gene.names, no.samples,
  effects, genes.pw, prop.de, seed = NULL)
}
\arguments{
\item{mvn}{Object of class \code{\link[Umpire]{MVN}} for efficient simulation
of data from a multivariate normal distribution (Note: either mvn or
cov.matrix need to be specified).}

\item{cov.matrix}{Matrix with covariances, as estimated with the
\code{\link[stats]{cov}} function (Note: either mvn or
cov.matrix need to be specified).}

\item{gene.names}{Vector with names of all genes that should be simulated
(Note: needs to have the same dimension as cov.matrix).}

\item{no.samples}{Total number of individuals to simulate. A balanced number
of cases and controls will be simulated.}

\item{effects}{Vector with effect sizes (mean differences) for differentially
expressed genes.}

\item{genes.pw}{Vector with names of genes in selected
pathway.}

\item{prop.de}{Proportion of genes in selected pathway to be differentially
expressed.}

\item{seed}{Random seed. Default is NULL, which generates the seed from R.}
}
\value{
List with the following components:
\itemize{
\item \code{data.train} data.frame of training data with samples in rows and
variables in columns; the first column, called y, denotes the case-control
status (0: control, 1: case)
\item \code{data.test} data.frame of test data with samples in rows and
variables in columns; the first column, called y, denotes the case-control
status (0: control, 1: case)
\item \code{info.sim} named vector with simulated effect sizes for each of
the genes in the selected pathway (Note: list of genes might have been
reduced to the genes available in gene.names)
}
}
\description{
Simulates pathway based gene expression data using given
covariance matrix. A proportion of genes of a single pathway is randomly
selected and assigned a random selection of given effect sizes.
}
\details{
Randomly selected effect sizes are assigned to a random subset of
the genes in the selected pathway. Two data sets for this scenario are then
simulated and denoted as training and test data. This function uses the
\href{https://cran.r-project.org/web/packages/Umpire/index.html}{Umpire}
package for efficient simulation of multivariate normally distributed
random variables.
}
\examples{
# set seed to make analysis reproducible
set.seed(42)

# define effect sizes of differentially expressed genes
effects = c(runif(10, -3, -0.5),
            runif(10, 0.5, 3))

# random covariance matrix for 100 genes
no.genes = 100
m = matrix(rnorm(no.genes * no.genes, sd = 0.5),
           nrow = no.genes)
cov.matrix = t(m) \%*\% m

# simulate one replicate with 30 samples and first 20 genes belonging to
# selected pathway
sim.data.l = sim.data.study.2(cov.matrix = cov.matrix,
                              gene.names = paste0("gene", seq_len(no.genes)),
                              genes.pw = paste0("gene", 1:20),
                              effects = effects,
                              prop.de = 0.5,
                              seed = 42,
                              no.samples = 30)

# alternatively use MVN object (might be more efficient if many replicates
# are simulated)
mvn = Umpire::MVN(mu = rep(0, nrow(cov.matrix)),
                  Sigma = cov.matrix, tol = 1e-06)
sim.data.l = sim.data.study.2(mvn = mvn,
                              gene.names = paste0("gene", seq_len(no.genes)),
                              genes.pw = paste0("gene", 1:20),
                              effects = effects,
                              prop.de = 0.5,
                              seed = 42,
                              no.samples = 30)

}
\references{
Zhang J, Roebuck PL, Coombes KR. (2012) Simulating gene expression data to
estimate sample size for class and biomarker discovery. Int J Adv Life Sci.
4:44–51. \url{https://doi.org/10.1186/1471-2105-13-S13-S1}.
}
