% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pw_rf_pred_error.R
\name{pw.rf.pred.error}
\alias{pw.rf.pred.error}
\title{Pathway-guided random forest based on prediction error}
\usage{
pw.rf.pred.error(x = NULL, y = NULL, info.pw = NULL,
  type = "regression", min.no.var = 5, not.assoc.pw = TRUE,
  p.adjust.method = "BH", no.perm = 20, ...)
}
\arguments{
\item{x}{Matrix or data.frame of predictor variables with variables in
columns and samples in rows (Note: missing values are not allowed).}

\item{y}{Vector with values of outcome variable (Note: will be converted
to factor if classification mode is used).}

\item{info.pw}{Matrix or data.frame assigning variables to pathways with
variables in rows and pathways in columns (0: variable does not belong to
pathway, 1: variable belongs to pathway).}

\item{type}{Type of random forest ('regression' or 'classification'). Default
is 'regression'.}

\item{min.no.var}{Minimal number of variables per pathway (Note: pathways
with less variables are removed or moved into 'not_associated_genes_pw').
Default is 5.}

\item{not.assoc.pw}{Combines all variables not associated to any pathway
or pathways with less then <min.no.var> into pathway
'not_associated_genes_pw'. Default is TRUE.}

\item{p.adjust.method}{Method for multiple testing adjusting (see
\code{\link{p.adjust}}). Default is "BH" which is different from the
adjustment method used prior to version 0.4.0 ("bonferroni").}

\item{no.perm}{Number of permutations used to determine the P value. Default
is 20.}

\item{...}{Additional parameters used in the ranger or the associated
wrapper function.}
}
\value{
List with the following components:
\itemize{
\item \code{results.pw} data.frame with information about each pathway
\itemize{
\item \code{id} pathway identifier
\item \code{no.var} number of variables in pathway
\item \code{pred.error} prediction error of pathway specific RF
\item \code{pred.error.perm.min}, \code{pred.error.perm.median},
\code{pred.error.perm.max} summary information about prediction errors after
permuting the outcome
\item \code{pval} permutation based P value
\item \code{pval.adj} adjusted permutation based P value
\item \code{selected} 0 or 1 denoting if pathway is selected
}
\item \code{pw.sel} vector with ids of selected pathways
\item \code{info.pw} actually used variables to pathways assignment
}
}
\description{
Trains a separate random forest (RF) for each pathway and
selects pathways with out-of-bag (OOB) prediction error significantly lower
than expected by chance using a permutation approach proposed by Hedinger et
al. (2019). OOB prediction errors are calculated for RFs trained on data sets
with permuted outcome and the mean and standard deviation of all those
prediction errors (across pathways and repetitions) are estimated. The
P values are then derived using a normal distribution. For
more details about random forests parameters see \code{\link{wrapper.rf}} and
for more details about the method see Pang et al. (2006).
}
\details{
The method specific parameter \code{no.perm} can be adapted.
However, it is recommended to use a sufficiently large number so that
no.perm * number of pathways > 1000.
}
\examples{
# define pathway parameters
info.par = data.frame(pw = paste0("pw", 1:3),
                      no.genes = rep(20, 3),
                      prop.de = c(0, 0.5, 1),
                      cor = rep(0.6, 3),
                      stringsAsFactors = FALSE)

# simulate 30 samples (including 15 cases)
sim.data.l = sim.data.study.1(info.par = info.par,
                              no.samples = 30,
                              no.cases = 15,
                              seed = 42,
                              gamma.range = c(0.5, 1.5))

# Note: no.perm of 20 only used for illustration (should be larger for
# analysis of a small number of pathways)
res.pe = pw.rf.pred.error(x = sim.data.l$data.train[, -1],
                          y = sim.data.l$data.train[, 1],
                          info.pw = sim.data.l$info.pw,
                          type = "classification",
                          not.assoc.pw = FALSE,
                          no.perm = 20)

}
\references{
\itemize{
\item Hediger, S et al. (2019) On the use of random forest for two-sample testing.
arXiv. \url{https://arxiv.org/abs/1903.06287v2}
\item Pang, H et al. (2006) Pathway analysis using random forests
classification and regression. Bioinformatics 22: 2028-2036.
\url{https://doi.org/10.1093/bioinformatics/btl344}.
}
}
