% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper_rf.R
\name{wrapper.rf}
\alias{wrapper.rf}
\title{Wrapper function to call ranger}
\usage{
wrapper.rf(x = NULL, y = NULL, num.trees = 500, mtry.prop = NULL,
  type = "regression", importance = "impurity_corrected",
  seed = NULL, ...)
}
\arguments{
\item{x}{Matrix or data.frame of predictor variables with variables in
columns and samples in rows (Note: missing values are not allowed).}

\item{y}{Vector with values of outcome variable (Note: will be converted
to factor if classification mode is used).}

\item{num.trees}{Number of trees. Default is 500.}

\item{mtry.prop}{Proportion of variables that should be used at each split.
Default value is the number of variables^(3/4) as recommended by
Ishwaran et al. (2011) for minimal depth importance. (Note: This is a different
default value than in the \code{\link[ranger]{ranger}} function).}

\item{type}{Type of random forest ('regression' or 'classification'). Default
is 'regression'.}

\item{importance}{Variable importance mode ('none', 'impurity',
'impurity_corrected' or 'permutation'). Default is 'impurity_corrected'.}

\item{seed}{Random seed. Default is NULL, which generates the seed from R.
Set to 0 to ignore the R seed.}

\item{...}{Additional parameters used in the ranger function.}
}
\value{
Object of class ranger
}
\description{
Provides an interface to the random forest algorithm implemented
in \code{\link[ranger]{ranger}}. For more details about random forests
parameters see \code{\link[ranger]{ranger}}.
}
\references{
\itemize{
  \item Ishwaran H, Kogalur UB, Chen X, Minn AJ (2011). Random survival
  forests for high-dimensional data. Stat Anal Data Min 4:115–32.
  \url{ https://doi.org/10.1002/sam.10103}.
  \item Wright, MN & Ziegler, A (2017). ranger: A fast implementation of
  random forests for high dimensional data in C++ and R. J Stat Softw
  77: 1-17. \url{https://doi.org/10.18637/jss.v077.i01}.
  }
}
